move maxSigned31BitInt.js
