+++ packages/react-reconciler/src/ReactFiberUnwindWork.js
@@ -64,7 +64,7 @@ import {
 import {Sync} from './ReactFiberExpirationTime';
 
 import invariant from 'shared/invariant';
-import maxSigned31BitInt from './maxSigned31BitInt';
+import maxSigned31BitInt from 'shared/maxSigned31BitInt';
 import {
   expirationTimeToMs,
   LOW_PRIORITY_EXPIRATION,
